+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Refactored constraint functions into different files
 - Added "Total load (%)" statistic to output schema
 - Updated Fault currents schema to match faults input schema
+- Updated fault problem to prune buses that are missing from problem to avoid data issues
 - Tuned default solver instance settings for new problem definitions
 - Added helper function `build_settings_file` to aid in building a settings file specific to a network
 - Added helper function `build_events_file` to aid in building a (very) simple events file specific to a network